app/requirements.txt
cffi==2.0.0
cryptography==45.0.7
fdb==2.0.2
future==1.0.0
oracledb==3.3.0
pycparser==2.23
python-dotenv==1.1.1
